Added Protected Route Middleware

+++ backend/controller/goalController.js
@@ -1,11 +1,11 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
-
+const User = require("../models/userModel");
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
     res.status(200).json({
         status: "success",
         result: goals.length,
@@ -31,6 +31,7 @@ const setGoal = asyncHandler(async (req, res) => {
     }
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id,
     });
     res.status(200).json({ status: "success", data: { goal } });
 });
@@ -40,9 +41,21 @@ const setGoal = asyncHandler(async (req, res) => {
 // @access  Private
 const updateGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id);
+
     if (!goal) {
         res.status(400);
         throw new Error("Goal Not Found");
+    }
+    const user = await User.findById(req.user.id);
+    // Check User
+    if (!user) {
+        res.status(401);
+        throw new Error("User Not Found");
+    }
+    // Logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("User not Authorized");
     }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -54,6 +67,24 @@ const updateGoal = asyncHandler(async (req, res) => {
 // @route   DELETE /api/goals/:id
 // @access  Private
 const deleteGoal = asyncHandler(async (req, res) => {
+    const goal = await Goal.findById(req.params.id);
+
+    if (!goal) {
+        res.status(400);
+        throw new Error("Goal Not Found");
+    }
+    const user = await User.findById(req.user.id);
+    // Check User
+    if (!user) {
+        res.status(401);
+        throw new Error("User Not Found");
+    }
+    // Logged in user matches the goal user
+    console.log(user._id, user.id);
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("User not Authorized");
+    }
     await Goal.findByIdAndDelete(req.params.id);
     res.status(200).json({ id: req.params.id });
 });
